+++ familyhelper-ui-view/familyhelper-ui-view-pc/src/views/items/assetsManagement/assetBom/FileTabPanel.vue
@@ -145,10 +145,8 @@
     />
     <file-create-dialog
       title="新建文件"
-      type="ITEM"
       :visible.sync="createDialog.visible"
-      :parent-id="itemId"
-      @onItemFileChanged="handleSearch"
+      @onConfirmed="handleCreateConfirmed"
     />
   </div>
 </template>
@@ -448,6 +446,28 @@ export default {
           this.$message({
             showClose: true,
             message: '文件上传成功',
+            type: 'success',
+            center: true,
+          });
+        })
+        .then(() => {
+          this.handleSearch();
+        })
+        .then(() => {
+          callback(true);
+        })
+        .catch(() => {
+          callback(false);
+        });
+    },
+    handleCreateConfirmed(file, callback) {
+      const formData = new FormData();
+      formData.append('file', file.blob, file.name);
+      resolveResponse(upload(this.itemId, formData))
+        .then(() => {
+          this.$message({
+            showClose: true,
+            message: '文件新建成功',
             type: 'success',
             center: true,
           });